requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
Django==3.2.5
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
django-model-utils==4.1.0
django-graphql-auth
django-graphql-jwt
django-debug-toolbar
